+++ jellED/core/beatDetection/include/envelopeDetector.h
@@ -7,25 +7,25 @@ namespace jellED {
 
 class EnvelopeDetector : public FilterStage {
 public:
-    static constexpr double DEFAULT_ATTACK_TIME = 0.005;   // 5ms attack - fast enough to catch transients
-    static constexpr double DEFAULT_RELEASE_TIME = 0.050;  // 50ms release - smooth decay between beats
+    static constexpr float DEFAULT_ATTACK_TIME = 0.005f;   // 5ms attack - fast enough to catch transients
+    static constexpr float DEFAULT_RELEASE_TIME = 0.050f;  // 50ms release - smooth decay between beats
 
-    EnvelopeDetector(uint32_t sample_rate, uint8_t downsample_factor = 1, double envelope_gain = 1.0,
-                     double attack_time = DEFAULT_ATTACK_TIME, double release_time = DEFAULT_RELEASE_TIME);
+    EnvelopeDetector(uint32_t sample_rate, uint8_t downsample_factor = 1, float envelope_gain = 1.0f,
+                     float attack_time = DEFAULT_ATTACK_TIME, float release_time = DEFAULT_RELEASE_TIME);
     ~EnvelopeDetector();
-    double apply(const double sample);
+    float apply(const float sample);
 
-    void setTimings(double attackTime, double releaseTime);
+    void setTimings(float attackTime, float releaseTime);
 
 private:
-    double attack_coeff;
-    double release_coeff;
-    
+    float attack_coeff;
+    float release_coeff;
+
     uint32_t sample_rate;
     uint8_t downsample_factor;
-    double current_envelope;
+    float current_envelope;
     uint8_t sample_counter;
-    double envelope_gain;
+    float envelope_gain;
 };
 
 } // namespace jellED
