+++ jellED/core/beatDetection/src/envelopeDetector.cpp
@@ -5,24 +5,24 @@
 
 namespace jellED {
 
-EnvelopeDetector::EnvelopeDetector(uint32_t sample_rate, uint8_t downsample_factor, double envelope_gain,
-                                   double attack_time, double release_time)
+EnvelopeDetector::EnvelopeDetector(uint32_t sample_rate, uint8_t downsample_factor, float envelope_gain,
+                                   float attack_time, float release_time)
 : sample_rate{sample_rate},
   downsample_factor{downsample_factor},
-  current_envelope{0.0},
+  current_envelope{0.0f},
   sample_counter{0},
   envelope_gain{envelope_gain}
 {
-    this->attack_coeff = 1.0 - std::exp(-1.0 / (attack_time * this->sample_rate));
-    this->release_coeff = 1.0 - std::exp(-1.0 / (release_time * this->sample_rate));
+    this->attack_coeff = 1.0f - std::exp(-1.0f / (attack_time * this->sample_rate));
+    this->release_coeff = 1.0f - std::exp(-1.0f / (release_time * this->sample_rate));
 }
 
 EnvelopeDetector::~EnvelopeDetector() {
 }
 
-double EnvelopeDetector::apply(const double sample) {
+float EnvelopeDetector::apply(const float sample) {
     // 1. Rectify the sample
-    double sample_abs = std::abs(sample);
+    float sample_abs = std::abs(sample);
 
     // 2. Smooth with single pole lowpass filter (attack/release envelope follower)
     if (sample_abs > current_envelope) {
@@ -34,7 +34,7 @@ double EnvelopeDetector::apply(const double sample) {
     // 3. Downsample
     sample_counter++;
     if (sample_counter % downsample_factor != 0) {
-        return -1.0;
+        return -1.0f;
     }
 
     // 4. Return the envelope value directly (scaled)
@@ -43,9 +43,9 @@ double EnvelopeDetector::apply(const double sample) {
     return current_envelope * envelope_gain;
 }
 
-void EnvelopeDetector::setTimings(double attackTime, double releaseTime) {
-    attack_coeff = 1.0 - std::exp(-1.0 / (attackTime * sample_rate));
-    release_coeff = 1.0 - std::exp(-1.0 / (releaseTime * sample_rate));
+void EnvelopeDetector::setTimings(float attackTime, float releaseTime) {
+    attack_coeff = 1.0f - std::exp(-1.0f / (attackTime * sample_rate));
+    release_coeff = 1.0f - std::exp(-1.0f / (releaseTime * sample_rate));
 }
 
 } // end namespace jellED
