+++ jellED/core/beatDetection/src/ringbuffer.cpp
@@ -15,13 +15,13 @@ Ringbuffer::Ringbuffer(uint32_t max_len, bool initializeScratchBuffer)
   max_len{max_len},
   current_size{0}
  {
-    buffer = (double*) malloc(sizeof(double) * max_len);
+    buffer = (float*) malloc(sizeof(float) * max_len);
     tail = &buffer[max_len - 1];
     for (uint32_t index = 0; index < max_len; ++index) {
-        this->append(0.0);
+        this->append(0.0f);
     }
     if (initializeScratchBuffer) {
-        scratch_buffer = (double*) malloc(sizeof(double) * max_len);
+        scratch_buffer = (float*) malloc(sizeof(float) * max_len);
     } else {
         scratch_buffer = nullptr;
     }
@@ -49,17 +49,17 @@ uint32_t Ringbuffer::get_position(const uint32_t index) const {
     }
 }
 
-void Ringbuffer::fill(double value) {
+void Ringbuffer::fill(float value) {
     for (uint32_t index = 0; index < max_len; ++index) {
         this->append(value);
     }
 }
 
-double Ringbuffer::get(uint32_t index) const {
+float Ringbuffer::get(uint32_t index) const {
     return this->buffer[get_position(index)];
 }
 
-void Ringbuffer::append(double entry) {
+void Ringbuffer::append(float entry) {
     this->head_position = (this->head_position + 1) % this->max_len;
     uint32_t tail_pos = (this->head_position + this->max_len - 1) % this->max_len;
     this->tail = &buffer[tail_pos];
@@ -70,7 +70,7 @@ void Ringbuffer::append(double entry) {
     }
 }
 
-void Ringbuffer::override_head_value(double value) {
+void Ringbuffer::override_head_value(float value) {
     //this->buffer[this->head_position] = value;
     *this->tail = value;
 }
@@ -79,7 +79,7 @@ void Ringbuffer::clear() {
     this->head_position = 0;
     this->current_size = 0;
     for (uint32_t index = 0; index < max_len; ++index) {
-        this->buffer[index] = 0.0;
+        this->buffer[index] = 0.0f;
     }
 }
 
@@ -87,11 +87,11 @@ uint32_t Ringbuffer::size() const {
     return current_size;
 }
 
-double Ringbuffer::min() const {
+float Ringbuffer::min() const {
     if (current_size == 0) {
-        return 0.0;
+        return 0.0f;
     }
-    double min_val = std::numeric_limits<double>::max();
+    float min_val = std::numeric_limits<float>::max();
     for (uint32_t i = 0; i < current_size; ++i) {
         uint32_t pos = get_position(i);
         if (buffer[pos] < min_val) {
@@ -101,11 +101,11 @@ double Ringbuffer::min() const {
     return min_val;
 }
 
-double Ringbuffer::max() const {
+float Ringbuffer::max() const {
     if (current_size == 0) {
-        return 0.0;
+        return 0.0f;
     }
-    double max_val = std::numeric_limits<double>::lowest();
+    float max_val = std::numeric_limits<float>::lowest();
     for (uint32_t i = 0; i < current_size; ++i) {
         uint32_t pos = get_position(i);
         if (buffer[pos] > max_val) {
@@ -115,26 +115,26 @@ double Ringbuffer::max() const {
     return max_val;
 }
 
-double Ringbuffer::median() const {
+float Ringbuffer::median() const {
     if (current_size == 0) {
-        return 0.0;
+        return 0.0f;
     }
-    
+
     // Copy to scratch buffer
     for (uint32_t i = 0; i < current_size; ++i) {
         scratch_buffer[i] = buffer[get_position(i)];
     }
-    
+
     // Partial sort using nth_element logic (or simple insertion sort for small N)
-    std::nth_element(scratch_buffer, scratch_buffer + current_size/2, 
+    std::nth_element(scratch_buffer, scratch_buffer + current_size/2,
                      scratch_buffer + current_size);
-    
-    double med = scratch_buffer[current_size / 2];
+
+    float med = scratch_buffer[current_size / 2];
     if (current_size % 2 == 0) {
         // For even count, average the two middle values
-        double max_lower = *std::max_element(scratch_buffer, 
-                                              scratch_buffer + current_size/2);
-        med = (med + max_lower) / 2.0;
+        float max_lower = *std::max_element(scratch_buffer,
+                                             scratch_buffer + current_size/2);
+        med = (med + max_lower) / 2.0f;
     }
     return med;
 }
